// ui/src/api/generated/models/default-parameter-mapping-info.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Halo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2.25.4
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface DefaultParameterMappingInfo
 */
export interface DefaultParameterMappingInfo {
    /**
     * 
     * @type {string}
     * @memberof DefaultParameterMappingInfo
     */
    'mode'?: string;
    /**
     * 
     * @type {string}
     * @memberof DefaultParameterMappingInfo
     */
    'template'?: string;
}