// ui/src/api/generated/models/selection.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Halo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2.25.4
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ReasoningMapping } from './reasoning-mapping';

/**
 * 
 * @export
 * @interface Selection
 */
export interface Selection {
    /**
     * Optional constrained native field/path override
     * @type {string}
     * @memberof Selection
     */
    'field'?: string;
    /**
     * Inheritance or override mode
     * @type {string}
     * @memberof Selection
     */
    'mode'?: SelectionModeEnum;
    /**
     * 
     * @type {ReasoningMapping}
     * @memberof Selection
     */
    'reasoningMapping'?: ReasoningMapping;
    /**
     * Registered template id; required only for TEMPLATE mode
     * @type {string}
     * @memberof Selection
     */
    'template'?: string;
}

export const SelectionModeEnum = {
    Inherit: 'INHERIT',
    Template: 'TEMPLATE',
    Unsupported: 'UNSUPPORTED'
} as const;

export type SelectionModeEnum = typeof SelectionModeEnum[keyof typeof SelectionModeEnum];